+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2019-09-09 – 2019-09-15
+- Add occupancy diagram
 - Correct many, many small mistakes in the text
 - Try to remove code from the infiltrator demo but it is not working